requirements.txt
# Core Dependencies
streamlit==1.39.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1

# Web Scraping
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.2.2

# AI/ML Dependencies - Updated for compatibility
openai==1.99.4
langchain==0.2.12
langchain-community==0.2.11
langchain-openai==0.3.28
langchain-core==0.3.74
langchain-text-splitters==0.2.4
faiss-cpu==1.8.0.post1
tiktoken==0.7.0
tenacity==8.2.2
jiter==0.10.0

# Data Visualization
plotly==5.24.1
matplotlib==3.9.2
seaborn==0.13.2

# Utilities
tqdm==4.66.5
typing-extensions==4.12.2
pydantic==2.8.2
aiohttp==3.10.5

# Development (optional)
pytest==8.3.2
pytest-cov==5.0.0
black==24.8.0
flake8==7.1.1
ipython==8.27.0
